Fixes disabling privacy mode.
This appears to be a long-standing issue. Disabling audio and
restarting the thread removes the implementation, but not its
reference.

To restart, must first stop the audio, which supposedly didn't
exist, because it's in privacy mode, but because at some point
the audio thread existed and its reference still exists, the
main thread attempts to stop a thread that no longer exists.

So, this results in a deadlock in the thread_join.